requirements.txt
flask
openai
Pillow
telethon
python-dotenv
requests